[refactor]: Renamed folder and update namespaces
